Separated the handlebar setters from the service #5789
